+++ oteltest/config.go
@@ -14,49 +14,15 @@
 
 package oteltest // import "go.opentelemetry.io/otel/oteltest"
 
-import (
-	"context"
-	"encoding/binary"
-	"sync"
-	"sync/atomic"
-
-	"go.opentelemetry.io/otel/trace"
-)
-
-// defaultSpanContextFunc returns the default SpanContextFunc.
-func defaultSpanContextFunc() func(context.Context) trace.SpanContext {
-	var traceID, spanID uint64 = 1, 1
-	return func(ctx context.Context) trace.SpanContext {
-		var sc trace.SpanContext
-		if current := trace.SpanContextFromContext(ctx); current.IsValid() {
-			sc = current
-		} else {
-			var tid trace.TraceID
-			binary.BigEndian.PutUint64(tid[:], atomic.AddUint64(&traceID, 1))
-			sc = sc.WithTraceID(tid)
-		}
-		var sid trace.SpanID
-		binary.BigEndian.PutUint64(sid[:], atomic.AddUint64(&spanID, 1))
-		return sc.WithSpanID(sid)
-	}
-}
-
 type config struct {
-	// SpanContextFunc returns a SpanContext from an parent Context for a
-	// new span.
-	SpanContextFunc func(context.Context) trace.SpanContext
-
 	// SpanRecorder keeps track of spans.
 	SpanRecorder *SpanRecorder
 }
 
-func newConfig(opts ...Option) config {
+func newConfig(opts []Option) config {
 	conf := config{}
 	for _, opt := range opts {
 		opt.apply(&conf)
-	}
-	if conf.SpanContextFunc == nil {
-		conf.SpanContextFunc = defaultSpanContextFunc()
 	}
 	return conf
 }
@@ -64,20 +30,6 @@ func newConfig(opts ...Option) config {
 // Option applies an option to a config.
 type Option interface {
 	apply(*config)
-}
-
-type spanContextFuncOption struct {
-	SpanContextFunc func(context.Context) trace.SpanContext
-}
-
-func (o spanContextFuncOption) apply(c *config) {
-	c.SpanContextFunc = o.SpanContextFunc
-}
-
-// WithSpanContextFunc sets the SpanContextFunc used to generate a new Spans
-// context from a parent SpanContext.
-func WithSpanContextFunc(f func(context.Context) trace.SpanContext) Option {
-	return spanContextFuncOption{SpanContextFunc: f}
 }
 
 type spanRecorderOption struct {
@@ -92,50 +44,4 @@ func (o spanRecorderOption) apply(c *config) {
 // testing.
 func WithSpanRecorder(sr *SpanRecorder) Option {
 	return spanRecorderOption{SpanRecorder: sr}
-}
-
-// SpanRecorder performs operations to record a span as it starts and ends.
-// It is designed to be concurrent safe and can by used by multiple goroutines.
-type SpanRecorder struct {
-	startedMu sync.RWMutex
-	started   []*Span
-
-	doneMu sync.RWMutex
-	done   []*Span
-}
-
-// OnStart records span as started.
-func (ssr *SpanRecorder) OnStart(span *Span) {
-	ssr.startedMu.Lock()
-	defer ssr.startedMu.Unlock()
-	ssr.started = append(ssr.started, span)
-}
-
-// OnEnd records span as completed.
-func (ssr *SpanRecorder) OnEnd(span *Span) {
-	ssr.doneMu.Lock()
-	defer ssr.doneMu.Unlock()
-	ssr.done = append(ssr.done, span)
-}
-
-// Started returns a copy of all started Spans in the order they were started.
-func (ssr *SpanRecorder) Started() []*Span {
-	ssr.startedMu.RLock()
-	defer ssr.startedMu.RUnlock()
-	started := make([]*Span, len(ssr.started))
-	for i := range ssr.started {
-		started[i] = ssr.started[i]
-	}
-	return started
-}
-
-// Completed returns a copy of all ended Spans in the order they were ended.
-func (ssr *SpanRecorder) Completed() []*Span {
-	ssr.doneMu.RLock()
-	defer ssr.doneMu.RUnlock()
-	done := make([]*Span, len(ssr.done))
-	for i := range ssr.done {
-		done[i] = ssr.done[i]
-	}
-	return done
 }