+++ sdk/trace/sampling_test.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.opentelemetry.io/otel/sdk/trace/internal"
 
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/oteltest"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -241,7 +241,7 @@ func TestTracestateIsPassed(t *testing.T) {
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			traceState, err := oteltest.TraceStateFromKeyValues(attribute.String("k", "v"))
+			traceState, err := internal.TraceStateFromKeyValues(attribute.String("k", "v"))
 			if err != nil {
 				t.Error(err)
 			}